requirements.txt
transformers
datasets
evaluate
pillow
torch
torchvision
rouge_score
bleu
meteor